+++ src/protocol.h
@@ -26,10 +26,12 @@
 #ifndef PROTOCOL_H
 #define PROTOCOL_H
 
+#include <inttypes.h>
+
 struct Protocol {
     const char *const name;
-    const int default_port;
-    int (*const parse_packet)(const char*, size_t, char **);
+    const uint16_t default_port;
+    ssize_t (*const parse_packet)(const char*, size_t, char **);
     const char *const abort_message;
     const size_t abort_message_len;
 };
